added requests folder with file structure and test folder, implemented same strategy on rides folder and files with tests as well. Note - standard file structure for implementing DRY 2020. Imported requestsRouter into index. still need to double check on insomnia to double down on endpoint working correctly.
