+++ index.js
@@ -1,7 +1,9 @@
 const express = require("express");
 const authRouter = require("./auth/auth-router");
 const locationRouter = require("./location/location-router");
-const usersRouter = require("./user/user-router");
+const usersRouter = require("./users/users-router");
+//request router import added.
+const requestsRouter = require("./requests/requests-router");
 const PORT = process.env.PORT || 3001;
 const HOST = process.env.HOST || "localhost";
 const { verifyToken, validateUserToken } = require("./Middleware/auth");
@@ -27,6 +29,8 @@ app.get("/", function (req, res) {
 app.use("/auth", authRouter);
 app.use("/locations", verifyToken(), validateUserToken(), locationRouter);
 app.use("/users", verifyToken(), validateUserToken(), usersRouter);
+//request router added.
+app.use("/requests", requestsRouter);
 
 app.use((err, req, res, next) => {
     console.log(err);
