+++ requests/requests-post.js
@@ -1,0 +1,14 @@
+const express = require("express");
+const router = express.Router();
+const { Rides } = require("../Classes/rides");
+
+const rides = new Rides();
+
+router.post("/", async (req, res, next) => {
+    try {
+    } catch (err) {
+        next(err);
+    }
+});
+
+module.exports = router;
