+++ rides/__tests__/rides-post.spec.js
@@ -1,0 +1,11 @@
+// const supertest = require("supertest");
+// const server = require("../index");
+// const db = require("../database/db-config");
+// let ride;
+
+// beforeAll(async () => {
+//     await db.seed.run();
+//     rides = await (await supertest(server).post("/rides")).setEncoding
+//         start_location
+//     });
+// });
